database_systems/streamlit/requirements.txt
mysql-connector-python==8.0.31
streamlit==1.16.0
pandas~=1.5.2
numpy~=1.24.1
plotly~=5.11.0 # express and graph_objects
pycountry~=22.3.5